Scripts/Alfred.alfredpreferences/workflows/user.workflow.C8ADF3D7-1D26-405D-9607-07EEECCB1A78/requirements.txt
# Alfred Workflow Python Dependencies
# Alfred 工作流 Python 依赖

# Image processing - 图像处理
Pillow>=9.0.0

# HTTP requests - HTTP 请求
requests>=2.28.0

# XML processing - XML 处理 (optional for some features)
lxml>=4.9.0

# Additional utilities - 额外工具
urllib3>=1.26.0
certifi>=2022.0.0
